requirements.txt
boto3>=1.34.0
azure-identity>=1.15.0
azure-mgmt-storage>=21.0.0
azure-mgmt-authorization>=4.0.0
azure-mgmt-monitor>=6.0.0
azure-mgmt-network>=25.0.0
azure-graphrbac>=0.61.1
google-cloud-storage>=2.14.0
google-cloud-iam>=2.13.0
google-cloud-logging>=3.9.0
google-cloud-compute>=1.14.0
google-auth>=2.27.0
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
msgraph-sdk>=1.0.0
rich>=13.7.0
click>=8.1.7
jinja2>=3.1.3
